# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentacion API - Aulas",
        "description": "Listado de endpoints de producto - explicacion",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://apiaulas.test/"
        }
    ],
    "paths": {
...
        "/api/producto/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar la información de un solo producto",
                "operationId": "a6d49ccf55297bdf564ce000dad43f71",
                "parameters": [
                    {
                        "name": "id_producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_producto": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nom_producto": {
                                            "type": "string",
                                            "example": "Esfero Pelikan"
                                        },
                                        "precio_unitario": {
                                            "type": "number",
                                            "format": "double",
                                            "example": "1500"
                                        },
                                        "unidades_disponibles": {
                                            "type": "integer",
                                            "example": "30"
                                        },
                                        "marca": {
                                            "type": "string",
                                            "example": "Pelikan"
                                        },
                                        "proveedor_id_proveedor": {
                                            "type": "integer",
                                            "example": "2"
                                        },
                                        "categoria_producto": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\\\Models\\\\Cliente] #id"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizacion de un producto especificado",
                "operationId": "7f0b3e6c22284c29ebfc56e82b6bc213",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_producto": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "nom_producto": {
                                                        "type": "string",
                                                        "example": "Esfero Pelikan"
                                                    },
                                                    "precio_unitario": {
                                                        "type": "number",
                                                        "format": "double",
                                                        "example": "1500"
                                                    },
                                                    "unidades_disponibles": {
                                                        "type": "integer",
                                                        "example": "30"
                                                    },
                                                    "marca": {
                                                        "type": "string",
                                                        "example": "Pelikan"
                                                    },
                                                    "proveedor_id_proveedor": {
                                                        "type": "integer",
                                                        "example": "2"
                                                    },
                                                    "categoria_producto": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}